refactor: change /assets to /public
