+++ src/pages/MyHobbies.tsx
@@ -2,9 +2,9 @@
 import { motion } from "framer-motion";
 import Carousel from "react-material-ui-carousel";
 // images
-import InstagramPage from "../assets/MyHobbies/instagram-page.jpeg";
-import SteamPage from "../assets/MyHobbies/steam-page.jpeg";
-import NarakaRank from "../assets/MyHobbies/naraka-rank.jpg";
+import InstagramPage from "../public/MyHobbies/instagram-page.jpeg";
+import SteamPage from "../public/MyHobbies/steam-page.jpeg";
+import NarakaRank from "../public/MyHobbies/naraka-rank.jpg";
 // styled components
 import {
   BlackCenterTitle,
